backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.5.0
supabase==2.8.0
stripe==7.8.0
pandas==2.1.4
numpy==1.24.3
aiofiles==23.2.1
plaid-python==9.1.0
anthropic>=0.30.0
httpx==0.26.0
httpcore>=1.0.0
alpaca-py>=0.25.0
websocket-client>=1.6.0
python-dotenv>=1.0.0
apscheduler>=3.10.4
